+++ CHANGELOG.rst
@@ -1,6 +1,6 @@
 Unreleased
 ----------
-* 
+* Fixed a bug in ``get_sections`` when ``top_levels_only`` was ``True``.
 
 v0.56.0
 -------